feat: 방 삭제 로직 추가, 게임방 상태 저장

+++ src/main/java/com/example/web2_3_ourtuft_be/room/controller/LobbyController.java
@@ -1,6 +1,7 @@
 package com.example.web2_3_ourtuft_be.room.controller;
 
 import com.example.web2_3_ourtuft_be.global.response.GlobalResponse;
+import com.example.web2_3_ourtuft_be.redis.service.ParticipantService;
 import com.example.web2_3_ourtuft_be.room.dto.RoomRequestDto;
 import com.example.web2_3_ourtuft_be.room.dto.RoomResponseDto;
 import com.example.web2_3_ourtuft_be.room.service.LobbyService;
@@ -11,6 +12,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import java.util.List;
+import java.util.Map;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -24,6 +27,7 @@ import org.springframework.web.bind.annotation.*;
 @Tag(name = "📬 Lobby", description = "로비 및 방 관련 API")
 public class LobbyController {
     private final LobbyService lobbyService;
+    private final ParticipantService participantService;
 
     @Operation(summary = "방 전체 조회 API", description = "로비에서 생성된 방을 조회합니다.")
     @ApiResponses({
@@ -86,12 +90,30 @@ public class LobbyController {
         return ResponseEntity.ok(GlobalResponse.success("방장 변경 성공"));
     }
 
-    @Operation(summary = "방 삭제 API", description = "방을 삭제합니다.")
-    @DeleteMapping("/rooms/{roomId}")
-    public GlobalResponse<String> deleteRoom(@PathVariable Long roomId) {
+    @Operation(summary = "방 나가기 API", description = "사용자가 특정 방을 나갑니다. 잔여인원이 0명이 되면 방이 삭제됩니다.")
+    @DeleteMapping("/rooms/{roomId}/participant/{userId}")
+    public ResponseEntity<GlobalResponse<String>> leaveRoom(
+            @PathVariable Long roomId,
+            @PathVariable Long userId
+    ){
+        participantService.removeParticipant(roomId, userId);
 
-        lobbyService.deleteRoom(roomId);
+        // 남은 참가자수 확인
+        Map<String, String> participants = participantService.getParticipants(roomId);
 
-        return GlobalResponse.success("방이 삭제되었습니다.");
+        if(participants.isEmpty()) {
+            lobbyService.deleteRoom(roomId);
+            return ResponseEntity.ok(GlobalResponse.success("방이 삭제되었습니다."));
+
+        } else {
+            // 방장이 나갔을 경우 방장 변경
+            String newHostId = participantService.getNextHost(String.valueOf(roomId));
+            if(newHostId != null) {
+                lobbyService.changeRoomHost(roomId, Long.parseLong(newHostId));
+            }
+
+            return ResponseEntity.ok(GlobalResponse.success("방에서 나갔습니다."));
+        }
+
     }
 }
